Fixed pipeline

+++ caption_rating_module/caption_rating.py
@@ -78,7 +78,7 @@ class CaptionRating:
         caption_filter_csv = return_video_folder_name(self.video_runner_obj)+'/'+CAPTION_SCORE
         with open(caption_filter_csv, newline='', encoding='utf-8') as caption_filter_file:
             data = list(csv.DictReader(caption_filter_file))
-            filtered_list = [x['frame_index'] for x in data if float(x['rating']) > int(os.getenv('CAPTION_RATING_THRESHOLD'))]
+            filtered_list = [x['frame_index'] for x in data if float(x['rating']) > float(os.getenv('CAPTION_RATING_THRESHOLD'))]
 
         objcsvpath = return_video_folder_name(self.video_runner_obj)+'/'+OBJECTS_CSV
         with open(objcsvpath, newline='', encoding='utf-8') as objcsvfile:
